Control Signatures From Expression Bins

Return a control signature for the signature or <Group> provided. Control signatures are generated by binning the average row values (average gene expression values) and sampling from the resulting bins such that every value in the original signature has <n> bin-matched values in the resulting output.

binmatch(Group, mat = NULL, x = NULL, bins = NULL, nbin = 30,
  n = 100, replace = FALSE, ...)

Arguments

Group
a character vector.

mat
a matrix.

x
a named numeric vector to generate bins from. If NULL, x will be computed from mat. Default: NULL

bins
a named character vector with as names the rownames and as values the ID of each bin. Default: NULL

nbin
numeric value. How many bins should be created? Default: 30

n
numeric value for the number of control values to return per value in Group. Default: 100

replace
boolean value. Allow bin sampling to be done with replacement. Default: F

Value

a character vector of length n * length(Group), containing the binmatched values for the values (gene names) in Group.