Hierarchical Clustering and Differential Gene Expression

This function computes the clusters from the expression matrix specified by hierarchical clustering. It can also (1) filter the resulting clusters by size (if they pass min and max size cutoffs), (2) compute the sets of differentially expressed genes in each cluster, (3) filter the resulting clusters by gene significance (if they pass nsig1 and nsig2 cutoffs), (4) filter the resulting clusters by cluster similarity (if they pass the Jaccard similarity cutoff) such that for any pair of clusters with too-high similarity, the least-significant cluster is removed and (4) sort the resulting clusters by significance.

clusters(m = NULL, hc = NULL, clusters = NULL,
  cor.method = "pearson", hc.method = "average", compute.dist = T,
  dist.method = "euclidean", k = NULL, h = NULL, size.filter = F,
  min.size = T, max.size = T, min.rel = 0.05, max.rel = 0.8,
  min.abs = 5, DEA = T, FC = 2, is.log = T, p = 0.01,
  adjust.method = "BH", sort = T, p.sort = F, return.full = F,
  return.p = F, fast = T, significance.filter = F, nsig1 = 30,
  nsig2 = 5, similarity.filter = F, all.filters = T,
  similarity.cutoff = 0.8, order.method = "combined",
  ties.method = "random", first = "sig1", decreasing = TRUE,
  sortBySignificance = T, verbose = F)

Arguments

m
matrix. Default: NULL

hc
hierarchical clustering (hclust) object. If provided, will be used instead of computed. Default: NULL

cor.method
a character string indicating which correlation coefficient is to be computed. Default: 'pearson'

hc.method
a character string indicating which agglomeration method to use. Default: 'average'

compute.dist
a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T

dist.method
a character string specifying the distance metric to be used for hierarchical clustering. Default: 'euclidean'

k
number of clusters to return. Default: NULL

h
tree height at which to retrieve clusters. Default: NULL

size.filter
if TRUE, will return clusters that are within the sizes specified. Default: TRUE

min.size
a boolean value or numeric value indicating whether there should be a minimum number of clusters or what the minimum number of clusters should be, respectively. If the former, then the minimum number of clusters will be calculated relative to the total number of members with min.rel.

max.size
a boolean value or numeric value indicating whether there should be a maximum number of clusters or what the maximum number of clusters should be, respectively. If the former, then the maximum number of clusters will be calculated relative to the total number of members with max.rel.

min.rel
minimum relative size of a cluster. Default: 0.01 (ie. 0.01 of total number of members)

max.rel
maximum relative size of a cluster. Default: 0.8 (ie. 0.8 of total number of members)

min.abs
absolute minimum size. If min.rel < min.abs, min.abs will replace min.rel. Default: 5

DEA
a boolean value indicating whether to run Differential Gene Expression Analysis. Required if any of significance.filter or similarity.filter are set to TRUE. Default: T

FC
fold change cutoff. Passed to genes. Default: 2

is.log
values are in log2 (used in fold_changes). Passed to genes. Default: T

p
p-value cutoff. Passed to genes. Default: 0.01

adjust.method
Correction for multiple tests. If not desired, set to 'none'. Passed to genes. Default: 'BH'

p.sort
returning values are sorted by p-value (most significant first). Passed to genes. Default: F

return.full
return cluster membership and cluster DE genes. Passed to genes. Default: F

return.p
return cluster DE p-values. Passed to genes. Default: F

fast
if TRUE, computes p-values for significant fold-change values only. Passed to genes. Default: T

significance.filter
a boolean value indicating whether to remove clusters that do not pass gene significance cutoffs. Default: T

nsig1
cutoff for number of differentially expressed genes. Used to cut by significance (significance.filter = TRUE). If not desired, set to FALSE. Default: 30

nsig2
cutoff for number of most differentially expressed genes. Used to cut by significance (significance.filter = TRUE). If not desired, set to FALSE. Default: 5

similarity.filter
a boolean value indicating whether to remove clusters that are too similar. Default: T

all.filters
a boolean value indicating whether to set all filtering options to TRUE: size.filter, significance.filter, similarity.filter. Default: T

similarity.cutoff
numeric value between 0 and 1 defining the maximum Jaccard similarity allowed. Default: 0.8

order.method
character string indicating which ordering method should be used, one of 'combined', 'sequential', 'sig1' or 'sig2'. 'combined' orders by the mean of sig1 ranks and sig2 ranks. 'sequential' orders by sig1 and sig2 sequentially, ordering by the second only where there are ties in the first. 'sig1' orders by the number of elements in each vector. 'sig2' orders by the number of elements in each vector after filtering the list such that each element only appears once in the vector in which it's value is highest. Default: 'combined'

ties.method
character string specifying how ties are treated, see base::rank for more details.

first
character string indicating whether ordering should be first by sig1 and then by sig2 (default) or vice versa. One of 'sig1' or 'sig2'. Only relevant if order.method is 'sequential'. Default: 'sig1'

decreasing
boolean value indicating whether the list should be reordered by decreasing or increasing significance. If TRUE, the most significant vector will appear first. Default: TRUE

Value

with default parameters, will compute hierarchical clustering object of the pariwise correlations of the matrix provided and retrieve all clusters at all possible heights. If the user prefers a specified number of clusters or a specified height, these can instead be provided.

Details

DETAILS

See also

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}